requirements.txt
Flask
pydantic
requests
python-dotenv
boto3